gleanerViperVal no longer init'd at the root level. so create one for running tools.
